Add newer TS so we can use template types; implement in Util/Metadatable
