+++ package.json
@@ -44,6 +44,6 @@
     "coveralls": "^3.0.2",
     "mocha": "^5.2.0",
     "nyc": "^13.1.0",
-    "typescript": "^4.0.3"
+    "typescript": "^4.4.4"
   }
 }
